Changes based on reviewers comments
