+++ chartlets.js/src/lib/utils/isString.ts
@@ -1,0 +1,3 @@
+export function isString(signalName: unknown): signalName is string {
+  return typeof signalName === "string";
+}
